split mutations.ts and queries.ts to directories
