+++ frontend/src/APIClients/mutations/AuthMutations.ts
@@ -1,6 +1,5 @@
 import { gql } from "@apollo/client";
 
-// Auth
 export const LOGIN = gql`
   mutation Login($email: String!, $password: String!) {
     login(email: $email, password: $password) {
@@ -94,47 +93,3 @@ export const SIGNUP = gql`
     }
   }
 `;
-
-// Stories
-export const UPDATE_STORY_TRANSLATION_CONTENTS = gql`
-  mutation UpdateStoryTranslationContents(
-    $storyTranslationContents: [StoryTranslationContentRequestDTO]
-  ) {
-    updateStoryTranslationContents(
-      storyTranslationContents: $storyTranslationContents
-    ) {
-      story {
-        id
-      }
-    }
-  }
-`;
-
-export const CREATE_TRANSLATION = gql`
-  mutation CreateStoryTranslation(
-    $storyTranslationData: CreateStoryTranslationRequestDTO!
-  ) {
-    createStoryTranslation(storyTranslationData: $storyTranslationData) {
-      story {
-        id
-      }
-    }
-  }
-`;
-export type CreateTranslation = {
-  story: {
-    id: number;
-  };
-};
-
-export const ASSIGN_REVIEWER = gql`
-  mutation AssignUserAsReviewer($storyTranslationId: ID!, $userId: ID!) {
-    assignUserAsReviewer(
-      storyTranslationId: $storyTranslationId
-      userId: $userId
-    ) {
-      ok
-    }
-  }
-`;
-export type AssignReviewer = { ok: boolean };
