+++ frontend/src/APIClients/mutations/StoryMutations.ts
@@ -1,0 +1,44 @@
+import { gql } from "@apollo/client";
+
+export const UPDATE_STORY_TRANSLATION_CONTENTS = gql`
+  mutation UpdateStoryTranslationContents(
+    $storyTranslationContents: [StoryTranslationContentRequestDTO]
+  ) {
+    updateStoryTranslationContents(
+      storyTranslationContents: $storyTranslationContents
+    ) {
+      story {
+        id
+      }
+    }
+  }
+`;
+
+export const CREATE_TRANSLATION = gql`
+  mutation CreateStoryTranslation(
+    $storyTranslationData: CreateStoryTranslationRequestDTO!
+  ) {
+    createStoryTranslation(storyTranslationData: $storyTranslationData) {
+      story {
+        id
+      }
+    }
+  }
+`;
+export type CreateTranslationResponse = {
+  story: {
+    id: number;
+  };
+};
+
+export const ASSIGN_REVIEWER = gql`
+  mutation AssignUserAsReviewer($storyTranslationId: ID!, $userId: ID!) {
+    assignUserAsReviewer(
+      storyTranslationId: $storyTranslationId
+      userId: $userId
+    ) {
+      ok
+    }
+  }
+`;
+export type AssignReviewerResponse = { ok: boolean };
